+++ src/main.tsx
@@ -6,7 +6,6 @@ import App from "./App";
 import { initAppearance } from "./lib/appearance";
 import { initSounds } from "./lib/sounds";
 import { handleQuitRequested, loadBootWorkspace } from "./lib/appLifecycle";
-import { consumeInstalledUpdate } from "./lib/updateNotice";
 import "./index.css";
 
 initAppearance();
@@ -41,14 +40,12 @@ void listen("quit_requested", () => {
 
 void loadBootWorkspace().then(
   ({ windowTransfer, resumed, history, historyCwd }) => {
-    const installedUpdate = windowTransfer ? null : consumeInstalledUpdate();
     ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
       <React.StrictMode>
         <BootGate>
           <App
             windowTransfer={windowTransfer}
             resumed={resumed}
-            installedUpdate={installedUpdate}
             history={history}
             historyCwd={historyCwd}
           />